prevent reentracy during minting from messing up the data integrity (#1)
Move the _checkOnERC721Received to the end of the _safeMint function to prevent reentrancy from messing up the data integrity. The reentrancy can only happened after the minting process is completed or before the process starts.

However, it is highly recommended to implement reentrancy guard in your own mint function.